First pass at fixing some of the broken tests

+++ deeplearning4j-core/src/main/java/org/deeplearning4j/optimize/distance/CosineDistance.java
@@ -17,7 +17,7 @@ public class CosineDistance extends BaseDistanceFunction {
 
 	@Override
 	public Float apply(INDArray input) {
-		return (float) (1 - similarity.apply(input));
+    return 1 - similarity.apply(input);
 	}
 
 }
